requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-autoslug==1.9.9
django-countries==7.5.1
django-environ==0.10.0
django-filter==23.1
django-phonenumber-field==7.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
flake8==6.0.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
phonenumbers==8.13.10
Pillow==9.5.0
platformdirs==3.3.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
